Manifest: Use int for equality_ids in manifest schema per Iceberg spec (#3840) (#3842)
Write equality_ids as list<int> and keep reading legacy list<long> manifests (#3840)

The Iceberg spec defines data_file.equality_ids as list<int>, but PyIceberg
wrote list<long>, which strict readers such as iceberg-cpp reject. Fix the
manifest schema so new manifests follow the spec, and let the Avro resolver
read the legacy long element as int for that field only, so existing
manifests stay readable without loosening type promotion anywhere else.

Closes #3840.

+++ pyiceberg/manifest.py
@@ -295,7 +295,7 @@ DATA_FILE_TYPE: dict[int, StructType] = {
         NestedField(
             field_id=135,
             name="equality_ids",
-            field_type=ListType(element_id=136, element_type=LongType(), element_required=True),
+            field_type=ListType(element_id=136, element_type=IntegerType(), element_required=True),
             required=False,
             doc="Field ids used to determine row equality in equality delete files.",
         ),
@@ -390,7 +390,7 @@ DATA_FILE_TYPE: dict[int, StructType] = {
         NestedField(
             field_id=135,
             name="equality_ids",
-            field_type=ListType(element_id=136, element_type=LongType(), element_required=True),
+            field_type=ListType(element_id=136, element_type=IntegerType(), element_required=True),
             required=False,
             doc="Field ids used to determine row equality in equality delete files.",
         ),
